swaggerツールの導入とtimelineのIndexメソッドに例を記述
