+++ README.md
@@ -16,3 +16,27 @@ atlas migrate diff migration_name \
 atlas migrate apply \
     --dir "file://db/migrations" \
     --url "postgres://postgres:postgres@study_pal_db:5432/study_pal?search_path=public&sslmode=disable"
+
+
+# API仕様書の自動生成手順
+
+### controllerのメソッドに以下のようにして記述
+// timelines godoc
+//
+//	@Summary		タイムライン取得API
+//	@Description	タイムラインを取得します
+//	@Tags			timelines
+//	@Accept			json
+//	@Produce		json
+//	@Param			page_size		query		int		true "ページサイズ"
+//	@Param			prev_page_token	query		string	false "次のページのトークン"
+//	@Param			next_page_token	query		string	false "前のページのトークン"
+//	@Success		200				{object}	IndexResponse
+//	@Failure		400				{object}	app_types.ErrorResponse
+//	@Failure		500				{object}	app_types.ErrorResponse
+//	@Router			/timelines [get]
+func (t *TimelineController) Index(c *gin.Context)
+
+### swag fmtコマンドを実行してファイルを自動整形
+
+### swag initコマンドを実行してswaggerファイルを自動生成
