+++ app/app_types/error_response.go
@@ -1,0 +1,11 @@
+package app_types
+
+type ErrorResponse struct {
+	Errors []string `json:"errors"`
+}
+
+func NewErrorResponse(errors []string) *ErrorResponse {
+	return &ErrorResponse{
+		Errors: errors,
+	}
+}
